requirements.txt
google-genai
SpeechRecognition
pyttsx3
pyaudio
psutil
pyautogui
requests
pywhatkit
Pillow
